chore: move namespace and copyrights to clack and switch to ELv2
