+++ clack/channels/schemas.py
@@ -1,0 +1,7 @@
+"""
+Elastic License 2.0
+
+Copyright Clack and/or licensed to Clack under one
+or more contributor license agreements. Licensed under the Elastic License;
+you may not use this file except in compliance with the Elastic License.
+"""
