+++ clack/guilds/schemas.py
@@ -1,17 +1,9 @@
 """
-Copyright 2021-2022 twattle, Inc.
+Elastic License 2.0
 
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
+Copyright Clack and/or licensed to Clack under one
+or more contributor license agreements. Licensed under the Elastic License;
+you may not use this file except in compliance with the Elastic License.
 """
 from typing import Literal, TypedDict
 
@@ -20,7 +12,7 @@ from apiflask.fields import Boolean, DateTime, Field, Integer, List, Nested, Str
 from apiflask.validators import Length, OneOf
 from typing_extensions import NotRequired
 
-from twattle.users.schemas import PublicUserObject
+from clack.users.schemas import PublicUserObject
 
 
 class CreateGuild(Schema):
