+++ clack/json.py
@@ -1,0 +1,27 @@
+"""
+Elastic License 2.0
+
+Copyright Clack and/or licensed to Clack under one
+or more contributor license agreements. Licensed under the Elastic License;
+you may not use this file except in compliance with the Elastic License.
+"""
+import orjson
+
+
+class ORJSONDecoder:
+    def __init__(self, **kwargs):
+        # eventually take into consideration when deserializing
+        self.options = kwargs
+
+    def decode(self, obj):
+        return orjson.loads(obj)
+
+
+class ORJSONEncoder:
+    def __init__(self, **kwargs):
+        # eventually take into consideration when serializing
+        self.options = kwargs
+
+    def encode(self, obj):
+        # decode back to str, as orjson returns bytes
+        return orjson.dumps(obj).decode('utf-8')
