+++ create_user.py
@@ -1,7 +1,7 @@
 from dotenv import load_dotenv
 
-from twattle.database import connect, sync_tables
-from twattle.users import routes
+from clack.database import connect, sync_tables
+from clack.users import routes
 
 load_dotenv()
