+++ docker.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-sudo docker build -t twattle-api .
-sudo docker stop twattle-api
-sudo docker rm -v twattle-api
-sudo docker run --name twattle-api -p 5000:5000 -d twattle-api
+sudo docker build -t clack-api .
+sudo docker stop clack-api
+sudo docker rm -v clack-api
+sudo docker run --name clack-api -p 5000:5000 -d clack-api
