--- twattle/json.py
@@ -1,35 +1,0 @@
-"""
-Copyright 2021-2022 twattle, Inc.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-"""
-import orjson
-
-
-class ORJSONDecoder:
-    def __init__(self, **kwargs):
-        # eventually take into consideration when deserializing
-        self.options = kwargs
-
-    def decode(self, obj):
-        return orjson.loads(obj)
-
-
-class ORJSONEncoder:
-    def __init__(self, **kwargs):
-        # eventually take into consideration when serializing
-        self.options = kwargs
-
-    def encode(self, obj):
-        # decode back to str, as orjson returns bytes
-        return orjson.dumps(obj).decode('utf-8')
